+++ Source/passout/Private/Gameplay/FPSObjectiveActor.cpp
@@ -1,10 +1,10 @@
 // Fill out your copyright notice in the Description page of Project Settings.
 
 
-#include "FPSObjectiveActor.h"
+#include "Gameplay/FPSObjectiveActor.h"
 #include "Components/SphereComponent.h"
 #include "Kismet/GameplayStatics.h"
-#include "passout/passoutCharacter.h"
+#include "Gameplay/passoutCharacter.h"
 
 
 // Sets default values
