+++ Source/passout/Private/Gameplay/passoutCharacter.cpp
@@ -1,7 +1,7 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
-#include "passoutCharacter.h"
-#include "passoutProjectile.h"
+#include "Gameplay/passoutCharacter.h"
+#include "Gameplay/passoutProjectile.h"
 #include "Animation/AnimInstance.h"
 #include "Camera/CameraComponent.h"
 #include "Components/CapsuleComponent.h"
